Extend PtrVector unit test
- converted to catch2
- added test for isAvailable
